Simplify geometry in dhis2 import
